add a toggle to use a different component for custom names
DataComponentTypes.ITEM_NAME prevents renaming the item in an anvil.

DataComponentTypes.CUSTOM_NAME allows renaming the item in an anvil.

The boolean isFixed, added to withDisplayName defualts to false, however when set to true. A player can not rename it in an anvil which someone might want for whatever reason.